requirements.txt
Flask == 3.1.0
icecream == 2.1.4 
mysql-connector-python == 9.2.0
flask-session == 0.8.0
redis == 5.2.1
requests == 2.32.3